renamed a couple things
