+++ src/main/kotlin/net/teengamingnights/assemblymod/sql/tables/FactoryRecipesStore.kt
@@ -1,8 +1,8 @@
 package net.teengamingnights.assemblymod.sql.tables
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import net.teengamingnights.assemblymod.factory.IFactory
-import net.teengamingnights.assemblymod.factory.IRecipe
+import net.teengamingnights.assemblymod.factory.Factory
+import net.teengamingnights.assemblymod.factory.Recipe
 import net.teengamingnights.assemblymod.factory.recipes.Recipes
 import net.teengamingnights.assemblymod.sql.DataStore
 import net.teengamingnights.assemblymod.sql.get
@@ -30,7 +30,7 @@ object FactoryRecipesStore : DataStore {
         get() = Store
 
     @JvmStatic
-    fun getRecipes(factory: IFactory, callback: Consumer<List<IRecipe>>) {
+    fun getRecipes(factory: Factory, callback: Consumer<List<Recipe>>) {
         async {
             transaction {
                 Store.select {
@@ -43,7 +43,7 @@ object FactoryRecipesStore : DataStore {
         }
     }
 
-    fun clearRecipes(factory: IFactory) {
+    fun clearRecipes(factory: Factory) {
         launch {
             transaction {
                 Store.deleteWhere {
@@ -53,9 +53,9 @@ object FactoryRecipesStore : DataStore {
         }
     }
 
-    fun addRecipes(factory: IFactory, recipes: List<IRecipe>) {
+    fun addRecipes(factory: Factory, recipes: List<Recipe>) {
         launch {
-            recipes.map(IRecipe::getId).forEach { recipeId ->
+            recipes.map(Recipe::getId).forEach { recipeId ->
                 transaction {
                     Store.insert {
                         it[Store.factoryId] = factory.id
